+++ cmake/Modules/LocalLibyuv.cmake
@@ -1,4 +1,7 @@
-set(AVIF_LIBYUV_TAG "deeb764bb92b6305ee0cb5dae7a5940fdb457fa9")
+# When changing the commit below to a newer version of libyuv, it is best to make sure it is being used by chromium,
+# because the test suite of chromium provides additional test coverage of libyuv.
+# It can be looked up at https://source.chromium.org/chromium/chromium/src/+/main:DEPS?q=libyuv.
+set(AVIF_LIBYUV_TAG "6067afde563c3946eebd94f146b3824ab7a97a9c")
 
 set(AVIF_LIBYUV_BUILD_DIR "${AVIF_SOURCE_DIR}/ext/libyuv/build")
 # If ${ANDROID_ABI} is set, look for the library under that subdirectory.